Add values for agi/int crit conversions.
Previously placeholder values were used. Most values have only received
a cursory research.

+++ Character/Character.h
@@ -53,8 +53,8 @@ public:
     virtual int get_stamina_modifier() const = 0;
     virtual int get_intellect_modifier() const = 0;
     virtual int get_spirit_modifier() const = 0;
-    virtual int get_agi_needed_for_one_percent_phys_crit() const = 0;
-    virtual int get_int_needed_for_one_percent_spell_crit() const = 0;
+    virtual double get_agi_needed_for_one_percent_phys_crit() const = 0;
+    virtual double get_int_needed_for_one_percent_spell_crit() const = 0;
     virtual int get_ap_per_strength() const = 0;
     virtual int get_ap_per_agi() const = 0;
